python/requirements.txt
# Global Payments Pay by Link - Python Dependencies
# Core web framework and HTTP client
Flask==3.0.0
requests==2.31.0

# Environment configuration
python-dotenv==1.0.0

# Production WSGI server
gunicorn==21.2.0
